requirements.txt
pandas
sqlalchemy
pyodbc
psycopg2-binary
python-dotenv
pyarrow
fastparquet
hdfs
scikit-learn
joblib
shap
jupyterlab
requests
